Fixed previous problems (delete method, class names, proxy, test structure)
