--- service/src/main/java/com/vlad/repository/impl/DriverRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.vlad.repository.impl;
-
-import com.vlad.entity.Driver;
-import com.vlad.repository.BaseAbstractRepository;
-import jakarta.persistence.EntityManager;
-
-public class DriverRepositoryImpl extends BaseAbstractRepository<Long, Driver> {
-
-    public DriverRepositoryImpl(EntityManager entityManager) {
-        super(Driver.class, entityManager);
-    }
-}
