+++ service/src/main/java/com/vlad/repository/impl/TripRepository.java
@@ -1,0 +1,12 @@
+package com.vlad.repository.impl;
+
+import com.vlad.entity.Trip;
+import com.vlad.repository.AbstractRepository;
+import jakarta.persistence.EntityManager;
+
+public class TripRepository extends AbstractRepository<Long, Trip> {
+
+    public TripRepository(EntityManager entityManager) {
+        super(Trip.class, entityManager);
+    }
+}
