+++ service/src/main/java/com/vlad/util/HibernateUtil.java
@@ -1,6 +1,5 @@
 package com.vlad.util;
 
-import jakarta.persistence.EntityManagerFactory;
 import lombok.experimental.UtilityClass;
 import org.hibernate.SessionFactory;
 import org.hibernate.boot.model.naming.CamelCaseToUnderscoresNamingStrategy;
@@ -9,10 +8,10 @@ import org.hibernate.cfg.Configuration;
 @UtilityClass
 public class HibernateUtil {
 
-    public static EntityManagerFactory buildEntityManagerFactory() {
+    public static SessionFactory buildSessionFactory() {
         Configuration configuration = buildConfiguration();
         configuration.configure();
-        return configuration.buildSessionFactory().unwrap(EntityManagerFactory.class);
+        return configuration.buildSessionFactory();
     }
 
     public static Configuration buildConfiguration() {
