+++ service/src/test/java/com/vlad/util/HibernateTestUtil.java
@@ -1,6 +1,5 @@
 package com.vlad.util;
 
-import jakarta.persistence.EntityManagerFactory;
 import lombok.experimental.UtilityClass;
 import org.hibernate.SessionFactory;
 import org.hibernate.cfg.Configuration;
@@ -15,13 +14,12 @@ public class HibernateTestUtil {
         postgres.start();
     }
 
-    public static EntityManagerFactory buildEntityManagerFactory() {
+    public static SessionFactory buildSessionFactory() {
         Configuration configuration = HibernateUtil.buildConfiguration();
         configuration.setProperty("hibernate.connection.url", postgres.getJdbcUrl());
         configuration.setProperty("hibernate.connection.username", postgres.getUsername());
         configuration.setProperty("hibernate.connection.password", postgres.getPassword());
         configuration.configure();
-
-        return configuration.buildSessionFactory().unwrap(EntityManagerFactory.class);
+        return configuration.buildSessionFactory();
     }
 }
